Added failure recovery in the chance that a server or volume gets put into the 'error' state when a create request has been made
If servers or volumes enter 'error' state then the server or volume they are
paired with are rendered useless when doing the pairing in attachments. This
allows for successful pairings ie 'ACTIVE' servers and 'available' volumes to
be attached even though other members in the list might not be as lucky.